simplify summary files
